fix: upgrade package and rename component
