+++ src/components/CertificateDropZone/CertificateDropZone.test.tsx
@@ -8,7 +8,7 @@ import { CertificateDropZone } from "./CertificateDropZone";
 import { CertificateDropZoneContainer } from "./CertificateDropZoneContainer";
 import { MemoryRouter } from "react-router";
 import NetworkSectionModel from "../NetworkSection/NetworkSectionModel";
-import { HeaderIconState } from "../UI/Overlay/OverlayContent/Model";
+import { HeaderIconState } from "../UI/Overlay/OverlayContent/Modal";
 jest.mock("qr-scanner");
 
 const store = configureStore();
