+++ src/components/CertificateDropZone/CertificateDropZone.tsx
@@ -19,7 +19,7 @@ import { getChainId } from "../../utils/shared";
 import { View, ViewActionError, ViewVerificationError, ViewVerificationPending } from "../DocumentDropzone/Views";
 import { ViewTokenRegistryMismatch } from "../DocumentDropzone/Views/ViewTokenRegistryMismatch";
 import NetworkSectionModel from "../NetworkSection/NetworkSectionModel";
-import { HeaderIconState } from "../UI/Overlay/OverlayContent/Model";
+import { HeaderIconState } from "../UI/Overlay/OverlayContent/Modal";
 import { useNetworkSelect } from "./../../common/hooks/useNetworkSelect";
 import { LoadDemoCertificate } from "./LoadDemoCertificate";
